added admin system

+++ src/modules/product/product.service.ts
@@ -1,16 +1,22 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  Injectable,
+  NotFoundException,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { PrismaService } from 'src/database/PrismaService';
 import { CreateProductDTO } from './dto/create-product.dto';
 import { UpdateProductDTO } from './dto/update-product.dto';
+import { UpdateUserDTO } from '../user/dto/update-user.dto';
 
 @Injectable()
 export class ProductService {
   constructor(private prisma: PrismaService) {}
 
-  async create(data: CreateProductDTO) {
+  async create(data: CreateProductDTO, user: UpdateUserDTO) {
     data.ratings = 0;
     data.stars = 0;
     data.sold = 0;
+    data.vendor_id = user.id;
     return await this.prisma.product.create({
       data,
     });
@@ -30,25 +36,34 @@ export class ProductService {
     return productExists;
   }
 
-  async update(id: string, data: UpdateProductDTO) {
+  async update(id: string, data: UpdateProductDTO, user: UpdateUserDTO) {
     const productExists = await this.prisma.product.findUnique({
       where: { id },
     });
 
     if (!productExists) throw new NotFoundException("Product doesn't exists");
 
-    return await this.prisma.product.update({ data, where: { id } });
+    if (productExists.vendor_id === user.id || user.admin)
+      return await this.prisma.product.update({ data, where: { id } });
+    else
+      throw new UnauthorizedException(
+        'Only the vendor or an admin can update the product',
+      );
   }
 
-  async delete(id: string) {
+  async delete(id: string, user: UpdateUserDTO) {
     const productExists = await this.prisma.product.findUnique({
       where: { id },
     });
 
     if (!productExists) throw new NotFoundException("Product doesn't exists");
 
-    await this.prisma.product.delete({ where: { id } });
-
-    return { success: `Product '${productExists.title}' deleted` };
+    if (productExists.vendor_id === user.id || user.admin) {
+      await this.prisma.product.delete({ where: { id } });
+      return { success: `Product '${productExists.title}' deleted` };
+    } else
+      throw new UnauthorizedException(
+        'Only the vendor or an admin can delete the product',
+      );
   }
 }
